+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 ### Fixed
-- Resolved [#194](https://github.com/EmbarkStudios/cargo-about/issues/194) by returning a configuration error when a `clarify` entry specifies neither `files` nor `git`, instead of silently ignoring the clarification.
+- [PR#304](https://github.com/EmbarkStudios/cargo-about/pull/304) resolved [#194](https://github.com/EmbarkStudios/cargo-about/issues/194) by returning a configuration error when a `clarify` entry specifies neither `files` nor `git`, instead of silently ignoring the clarification.
+- [PR#306](https://github.com/EmbarkStudios/cargo-about/pull/306) resolved [#305](https://github.com/EmbarkStudios/cargo-about/issues/305) by removing the log warnings for using the deprecated fields related to clearly defined which was deprecated/removed in [PR#287](https://github.com/EmbarkStudios/cargo-about/pull/287).
 
 ## [0.9.0] - 2026-04-24
 ### Changed